// types.ts
export type TransactionType = 'credit' | 'debit' | 'pix' | 'boleto';

export interface Transaction {
  id: string;
  description: string;
  amount: number;
  date: string;
  bankName: string;
  category: string;
  type: TransactionType;
  icon: React.FC<{className?: string}>;
  notes?: string;
}